test: 책임의 변경으로 인한 테스트 케이스 위치 변경
